requirements.txt
google-cloud-bigquery>=3.0.1